Refactor: windowModal -> modal컴포넌트로 변경
