+++ FE/src/components/Modals/index.tsx
@@ -1,0 +1,3 @@
+export default function Modals() {
+  return <div>Modals</div>;
+}
