+++ FE/src/components/Modals/style.ts
@@ -1,17 +1,16 @@
 import styled from 'styled-components';
 
-export const ModalWrapper = styled.div`
+export const Overlay = styled.div`
   position: fixed;
-  text-align: center;
   left: 0;
   bottom: 0;
   top: 0;
   right: 0;
   z-index: 1022;
-  background-color: rgba(0, 0, 0, 0.3);
+  background-color: 'rgba(255, 255, 255, 0.75)';
 `;
 
-export const ModalContainer = styled.div`
+export const Content = styled.div`
   margin-top: 200px;
   display: inline-block;
   z-index: 1012;
@@ -20,7 +19,7 @@ export const ModalContainer = styled.div`
   user-select: none;
 `;
 
-export const CloseModalButton = styled.button`
+export const CloseButton = styled.button`
   position: absolute;
   right: 10px;
   top: 6px;
