+++ FE/src/components/Modals/useModal.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 
-import { ModalDispatchContext } from '@/components/Modal/Context';
+import { ModalDispatchContext } from '@/components/Modals/Context';
 
 export default function useModal(key: string) {
   const { open, close } = useContext(ModalDispatchContext);
